# docs/api/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "NISQ Analyzer API",
    "contact" : {
      "name" : "GitHub Repository",
      "url" : "https://github.com/UST-QuAntiL/nisq-analyzer"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.4.0"
  },
  "servers" : [ {
    "url" : "http://localhost:5010/nisq-analyzer",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/analysis-results/jobs/{resId}" : {
      "get" : {
        "tags" : [ "analysis-result" ],
        "description" : "Retrieve a single analysis job result",
        "operationId" : "getAnalysisJob",
        "parameters" : [ {
          "name" : "resId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AnalysisJobDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
    "/compiler-results/" : {
      "get" : {
        "tags" : [ "compiler-analysis-result" ],
        "description" : "Retrieve all compiler analysis results",
        "operationId" : "getCompilerAnalysisResults",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CompilerAnalysisResultListDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
    "/compiler-results/jobs/{resId}" : {
      "get" : {
        "tags" : [ "compiler-analysis-result" ],
        "description" : "Retrieve a single compilation result",
        "operationId" : "getCompilerAnalysisJob",
        "parameters" : [ {
          "name" : "resId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CompilationJobDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
    "/execution-results/" : {
      "get" : {
        "tags" : [ "execution-result" ],
        "description" : "Retrieve all execution results for an Implementation",
        "operationId" : "getExecutionResults",
        "parameters" : [ {
          "name" : "implId",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExecutionResultListDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/implementations/" : {
      "get" : {
        "tags" : [ "implementation" ],
        "description" : "Retrieve implementations for an algorithm",
        "operationId" : "getImplementations",
        "parameters" : [ {
          "name" : "algoId",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ImplementationListDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      },
      "post" : {
        "tags" : [ "implementation" ],
        "description" : "Create an implementation",
        "operationId" : "createImplementation",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ImplementationDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ImplementationDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        }
      }
    },
...
    "/qpu-selection-results/jobs/{resId}" : {
      "get" : {
        "tags" : [ "qpu-selection-result" ],
        "description" : "Retrieve a single QPU selection job",
        "operationId" : "getQpuSelectionJob",
        "parameters" : [ {
          "name" : "resId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "userId",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QpuSelectionJobDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
    "/selection-params" : {
      "get" : {
        "tags" : [ "root" ],
        "description" : "Retrieve selection parameters",
        "operationId" : "getSelectionParams",
        "parameters" : [ {
          "name" : "algoId",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ParameterListDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        }
      }
    },
    "/selection" : {
      "post" : {
        "tags" : [ "root" ],
        "description" : "Select implementations for an algorithm",
        "operationId" : "selectImplementations",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SelectionRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AnalysisJobDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/compiler-selection" : {
      "post" : {
        "tags" : [ "root" ],
        "description" : "Select the most suitable compiler for an implementation loaded from the given URL",
        "operationId" : "selectCompilerForFile_1",
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "$ref" : "#/components/schemas/compilerselection_body"
              }
            },
            "application/xml" : {
              "schema" : {
                "$ref" : "#/components/schemas/CompilerSelectionDto"
              }
            },
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CompilerSelectionDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CompilationJobDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/qpu-selection" : {
      "post" : {
        "tags" : [ "root" ],
        "description" : "Select the most suitable quantum computer for a quantum circuit loaded from the given URL",
        "operationId" : "selectQpuForCircuitFile_1",
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "$ref" : "#/components/schemas/qpuselection_body"
              }
            },
            "application/xml" : {
              "schema" : {
                "$ref" : "#/components/schemas/QpuSelectionDto"
              }
            },
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/QpuSelectionDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QpuSelectionJobDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/compilers" : {
      "get" : {
        "tags" : [ "root" ],
        "description" : "Retrieve compilers",
        "operationId" : "getCompilers",
        "parameters" : [ {
          "name" : "provider",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/hal+json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "AnalysisJobDto" : {
        "type" : "object",
        "properties" : {
          "analysisResultList" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AnalysisResultDto"
            }
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "implementedAlgorithm" : {
            "type" : "string",
            "format" : "uuid"
          },
          "time" : {
            "type" : "string",
            "format" : "date-time"
          },
          "inputParameters" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "ready" : {
            "type" : "boolean"
          },
          "initialMcdaJob" : {
            "type" : "string",
            "format" : "uuid"
          },
          "initialMcdaMethod" : {
            "type" : "string"
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        }
      },
...
      "CompilerAnalysisResultListDto" : {
        "type" : "object",
        "properties" : {
          "compilerAnalysisResultList" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CompilerAnalysisResultDto"
            }
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        }
      },
      "CompilationJobDto" : {
        "type" : "object",
        "properties" : {
          "compilerAnalysisResultList" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CompilerAnalysisResultDto"
            }
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "ready" : {
            "type" : "boolean"
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        }
      },
...
      "ExecutionResultListDto" : {
        "type" : "object",
        "properties" : {
          "executionResultDtos" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ExecutionResultDto"
            }
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        }
      },
      "ImplementationListDto" : {
        "type" : "object",
        "properties" : {
          "implementationDtos" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ImplementationDto"
            }
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        }
      },
      "QpuSelectionJobDto" : {
        "type" : "object",
        "properties" : {
          "qpuSelectionResultList" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/QpuSelectionResultDto"
            }
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "time" : {
            "type" : "string",
            "format" : "date-time"
          },
          "circuitName" : {
            "type" : "string"
          },
          "ready" : {
            "type" : "boolean"
          },
          "userId" : {
            "type" : "string"
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        }
      },
...
    }
  }
}